+++ index.css
@@ -16,6 +16,7 @@ body {
   padding-left: 55px;
   padding-right: 55px;
   padding-bottom: 10;
+  position: relative;
 }
 
 .title {
@@ -113,10 +114,48 @@ body {
   width: 550px;
 }
 
+.input-container-outer {
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+}
+
+.error-container {
+  visibility: hidden;
+  max-width: 300px;
+  position: absolute;
+  top: 40px;
+  right: 50px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border: 3px solid red;
+  border-radius: 10px;
+  padding: 10px;
+  color: red;
+  background-color: #2f3e53;
+  font-size: medium;
+  gap: 10px;
+  text-align: center;
+}
+
+.error-container h3 {
+  margin-top: 5px;
+  margin-bottom: 5px;
+}
+
+.error-container p {
+  margin-top: 0;
+  margin-bottom: 10px;
+}
+
 .input-container {
   display: flex;
+  flex-direction: row;
   justify-content: space-between;
-  width: 550px;
+  align-items: center;
+  width: 800px;
 }
 
 .input-container input {
