+++ index.html
@@ -12,6 +12,7 @@
             <h2 class="subtitle">Never use an insecure password again</h2>
             <button class="generate-passwords" >Generate Passwords</button>
             <hr class="horizontal-rule">
+          
                 <div class="input-container">
                     <div class="password-container">
                         <label for="password-1">Password 1</label>
@@ -28,9 +29,13 @@
                             <button class="copy-password-2">Copy</button>
                         </div>
                         <h3 class="copied-password" id="copied-password-2">Copied!</h3>
-
                     </div>
                 </div>
+                <div class="error-container" id="error-container">
+                    <h3>Error</h3>
+                    <p id="error"></p>
+                </div>
+ 
             <script src="index.js" type="module"></script>
         </div>
     </body>
